+++ api/applib/src/main/java/org/apache/isis/applib/services/sudo/SudoService.java
@@ -21,7 +21,7 @@ package org.apache.isis.applib.services.sudo;
 import java.util.concurrent.Callable;
 import java.util.function.UnaryOperator;
 
-import org.apache.isis.applib.services.iactn.ExecutionContext;
+import org.apache.isis.applib.services.iactnlayer.ExecutionContext;
 import org.apache.isis.applib.services.user.RoleMemento;
 import org.apache.isis.applib.services.user.UserService;
 import org.apache.isis.commons.functional.ThrowingRunnable;
