+++ core/interaction/src/main/java/org/apache/isis/core/interaction/session/InteractionLayer.java
@@ -18,7 +18,7 @@
  */
 package org.apache.isis.core.interaction.session;
 
-import org.apache.isis.applib.services.iactn.ExecutionContext;
+import org.apache.isis.applib.services.iactnlayer.ExecutionContext;
 import org.apache.isis.core.security.authentication.Authentication;
 
 import lombok.Getter;
