+++ core/runtimeservices/src/main/java/org/apache/isis/core/runtimeservices/session/AnonymousSession.java
@@ -18,7 +18,7 @@
  */
 package org.apache.isis.core.runtimeservices.session;
 
-import org.apache.isis.applib.services.iactn.ExecutionContext;
+import org.apache.isis.applib.services.iactnlayer.ExecutionContext;
 import org.apache.isis.applib.services.user.UserMemento;
 import org.apache.isis.core.security.authentication.AuthenticationAbstract;
