+++ core/security/src/main/java/org/apache/isis/core/security/authentication/singleuser/SingleUserAuthentication.java
@@ -18,7 +18,7 @@
  */
 package org.apache.isis.core.security.authentication.singleuser;
 
-import org.apache.isis.applib.services.iactn.ExecutionContext;
+import org.apache.isis.applib.services.iactnlayer.ExecutionContext;
 import org.apache.isis.applib.services.user.UserMemento;
 import org.apache.isis.core.security.authentication.AuthenticationAbstract;
